// frontend/src/api/models/createOrganisationRoleRequest.ts
/**
 * Generated by orval v6.8.1 🍺
 * Do not edit manually.
 * Service Release Manager API
 * OpenAPI spec version: v1
 */

export interface CreateOrganisationRoleRequest {
	organisationId: number;
	name: string;
	serviceWrite: boolean;
	serviceDelete: boolean;
	userRead: boolean;
	userWrite: boolean;
	userDelete: boolean;
}